Refactor page.js into modular components with improved mobile support
